Move FileCache to its own file
